refactor(agents): isolate tool schema projection leaf
